+++ src/inspect_ai/dataset/_sources/json.py
@@ -6,6 +6,7 @@ from typing import Any, cast
 
 import jsonlines
 
+from inspect_ai._util.asyncfiles import is_s3_filename
 from inspect_ai._util.file import file
 
 from .._dataset import (
@@ -29,7 +30,7 @@ def json_dataset(
     limit: int | None = None,
     encoding: str = "utf-8",
     name: str | None = None,
-    fs_options: dict[str, Any] = {},
+    fs_options: dict[str, Any] | None = None,
 ) -> Dataset:
     r"""Read dataset from a JSON file.
 
@@ -72,8 +73,12 @@ def json_dataset(
         else json_dataset_reader
     )
 
+    # use readahead cache by default for s3
+    if fs_options is None and is_s3_filename(json_file):
+        fs_options = dict(default_fill_cache=True, default_cache_type="readahead")
+
     # read and convert samples
-    with file(json_file, "r", encoding=encoding, fs_options=fs_options) as f:
+    with file(json_file, "r", encoding=encoding, fs_options=fs_options or {}) as f:
         name = name if name else Path(json_file).stem
         dataset = MemoryDataset(
             samples=data_to_samples(dataset_reader(f), data_to_sample, auto_id),